+++ src/clojure_plus/core.clj
@@ -1,0 +1,6 @@
+(ns clojure-plus.core)
+
+(defmacro defcmd [name & cmds]
+  `(do
+     (defn ~name [] ~@cmds)
+     (command-for ~(str name) ~name)))